deployment/security-agent/backend/lambda/requirements.txt
boto3==1.34.0
strands-agents==1.13.0
PyJWT==2.8.0